app_core

Module embeds

External embed support for rich content

This module provides support for embedding external content like YouTube videos, Spotify tracks/playlists, and generic link previews in posts and messages.

Structs

EmbedDetector

Embed detector for identifying and parsing embeddable URLs

LinkEmbed

Generic link embed

SpotifyEmbed

Spotify embed

YouTubeEmbed

YouTube video embed

Enums

EmbedError

Errors that can occur during embed operations

EmbedType

Type of external embed

ExternalEmbed

External embed that can be any supported type

SpotifyEmbedType

Spotify embed type

YouTubeThumbnailQuality

YouTube thumbnail quality

Type Aliases

Result

Result type for embed operations